feat(stdlib): extend Class A builtin coverage (assert, exit, panic, per-type print)
Add sleep_ms, exit, panic, assert/assert_eq/assert_ne, and the
per-type print/println monomorphic table to the HIR catalog. Add
the missing runtime symbols (hew_exit, hew_println_int/str/bool/f64)
to hew-runtime.

Acceptance fixtures in tests/v05-vertical-slice/accept/ cover each
new builtin end-to-end: assert.hew exits 0; assert_eq_fail.hew
exits non-zero with assertion message; exit_42.hew exits 42;
panic.hew exits 101; print_int/bool/f64.hew match expected stdout.

+++ hew-runtime/src/assert.rs
@@ -15,7 +15,7 @@ use std::os::raw::c_char;
 ///
 /// Called from compiled Hew programs via C ABI. No preconditions.
 #[no_mangle]
-pub unsafe extern "C" fn hew_assert(cond: i64) {
+pub unsafe extern "C" fn hew_assert(cond: u8) {
     if cond == 0 {
         eprintln!("assertion failed");
         // SAFETY: abort() is always safe to call.
@@ -133,7 +133,7 @@ pub unsafe extern "C" fn hew_assert_ne_f64(left: f64, right: f64) {
 ///
 /// Called from compiled Hew programs via C ABI. No preconditions.
 #[no_mangle]
-pub unsafe extern "C" fn hew_assert_eq_bool(left: i32, right: i32) {
+pub unsafe extern "C" fn hew_assert_eq_bool(left: u8, right: u8) {
     if left != right {
         let l = if left != 0 { "true" } else { "false" };
         let r = if right != 0 { "true" } else { "false" };
@@ -149,7 +149,7 @@ pub unsafe extern "C" fn hew_assert_eq_bool(left: i32, right: i32) {
 ///
 /// Called from compiled Hew programs via C ABI. No preconditions.
 #[no_mangle]
-pub unsafe extern "C" fn hew_assert_ne_bool(left: i32, right: i32) {
+pub unsafe extern "C" fn hew_assert_ne_bool(left: u8, right: u8) {
     if left == right {
         let l = if left != 0 { "true" } else { "false" };
         let r = if right != 0 { "true" } else { "false" };
